Enhancement: Extract auto-review test suite
