+++ test/AutoReview/SrcCodeTest.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2017 Andreas Möller.
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ *
+ * @see https://github.com/localheinz/classy
+ */
+
+namespace Localheinz\Classy\Test\AutoReview;
+
+use Localheinz\Test\Util\Helper;
+use PHPUnit\Framework;
+
+/**
+ * @internal
+ */
+final class SrcCodeTest extends Framework\TestCase
+{
+    use Helper;
+
+    public function testSrcClassesAreAbstractOrFinal(): void
+    {
+        $this->assertClassesAreAbstractOrFinal(__DIR__ . '/../../src');
+    }
+
+    public function testSrcClassesHaveTests(): void
+    {
+        $this->assertClassesHaveTests(
+            __DIR__ . '/../../src',
+            'Localheinz\\Classy\\',
+            'Localheinz\\Classy\\Test\\Unit\\'
+        );
+    }
+}
